+++ MSAL/src/public/MSALDefinitions.h
@@ -158,6 +158,27 @@ typedef NS_ENUM(NSUInteger, MSALDeviceMode)
 };
 
 /**
+ Platform SSO status on macOS device
+ */
+typedef NS_ENUM(NSUInteger, MSALPlatformSSOStatus)
+{
+    /*
+        Administrator hasn't configured Platform SSO in sso config.
+    */
+    MSALPlatformSSONotEnabled,
+    
+    /*
+     Administrator has configured Platform SSO in sso config. But device has not been registred with AAD via platform SSO
+     */
+    MSALPlatformSSOEnabledNotRegistered,
+    
+    /*
+     Administrator has configured Platform SSO in sso config and the device is registred with AAD via platform SSO
+     */
+    MSALPlatformSSOEnabledAndRegistered
+};
+
+/**
     The block that gets invoked after MSAL has finished getting a token silently or interactively.
     @param result       Represents information returned to the application after a successful interactive or silent token acquisition. See `MSALResult` for more information.
     @param error         Provides information about error that prevented MSAL from getting a token. See `MSALError` for possible errors.
@@ -172,7 +193,7 @@ typedef void (^MSALAccountsCompletionBlock)(NSArray<MSALAccount *> * _Nullable a
 /**
     The completion block that will be called when current account is loaded, or MSAL encountered an error.
  */
-typedef void (^MSALCurrentAccountCompletionBlock)(MSALAccount * _Nullable account, MSALAccount * _Nullable previousAccount, NSError * _Nullable error);
+typedef void (^MSALCurrentAccountCompletionBlock)(MSALAccount * _Nullable_result account, MSALAccount * _Nullable_result previousAccount, NSError * _Nullable error);
 
 /**
     The completion block that will be called when sign out is completed, or MSAL encountered an error.